% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_anova}
\alias{f_stat_anova}
\title{generate a datatframe with anova results from a data_ls list}
\usage{
f_stat_anova(data_ls, col_group, boxcox = F)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_group}{character vector denoting grouping variable}

\item{boxcox}{perform analysis on boxcox-transformed or numerical variables}
}
\value{
dataframe
}
\description{
returns a dataframe with shapiro, anova und kruskal p values
  supplemented with maximum difference of means and medians between groups
}
\examples{
df_anova = data_ls = f_clean_data(mtcars) \%>\%
  f_stat_anova('cyl')

df_anova
}
\seealso{
\code{\link[stringr]{str_c}}
\code{\link[purrr]{map}},\code{\link[purrr]{map_dbl}}
\code{\link{f_stat_anova}}
}
