% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_clean.R
\name{f_boxcox}
\alias{f_boxcox}
\title{f_boxcox}
\usage{
f_boxcox(data_ls)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}
}
\value{
returns a list
  \item{data}{the cleaned dataframe as tibble}
  \item{categoricals}{vector of column names containing categorical data}
  \item{numericals}{vector of column names containing numerical data}
  \item{ids}{vector of column names containing ids}
  \item{boxcox_names}{vector of column names containing boxcox transformed variables}
  \item{boxcox_data}{tibble containing boxcox transformed variables}
}
\description{
Takes a data_ls object generated by f_clean_data() and adds
  boxcox transformations of all numeric variables.
}
\details{
For a boxcox transformation all values mut be > 0. The function will
  automatically add the abs( min(x) ) + 0.0001 to all columns if they contain
  values <= 0
}
\examples{
data_ls = f_clean_data(mtcars)
f_manip_get_most_common_level( data_ls$data$cyl)
}
\seealso{
\code{\link{f_clean_data}}
}
