% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_clean.R
\name{f_clean_data_no_changes}
\alias{f_clean_data_no_changes}
\title{wrapper for f_clean_data without modifications to data}
\usage{
f_clean_data_no_changes(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
returns a list
  \item{data}{the cleaned dataframe as tibble}
  \item{categoricals}{vector of column names containing categorical data}
  \item{categoricals_ordered}{vector of column names containing all ordered categorical data}
  \item{numericals}{vector of column names containing numerical data}
  \item{ids}{vector of column names containing ids}
}
\seealso{
\code{\link{f_clean_data}}
}
