% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_content.R
\name{f_content}
\alias{f_content}
\title{opens an html index file for additional content}
\usage{
f_content(render_missing = F, overwrite = T)
}
\arguments{
\item{render_missing}{logical, Default: T}
}
\description{
opens a window in the default browser with links to all
  additional content files in oetteR, including vignettes and proof of
  concepts drafted as R and Rmd files. It includes the option of rendering
  all Rmd files into html. Which however can take around 30 minutes at least.
  You can find links to most of the additional content rendered as html on
  the github page of the package.
}
\examples{
\dontrun{
 f_vignettes()
}
}
