% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_content.R
\name{f_content_get_file_type}
\alias{f_content_get_file_type}
\title{returns files of specified type from vignette path}
\usage{
f_content_get_file_type(file_type = "html", folder = "vignettes")
}
\arguments{
\item{file_type}{character vector denoting file extensions}
}
\examples{
\dontrun{
 f_content_get_file_type()
}
}
