% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_content.R
\name{f_content_get_path}
\alias{f_content_get_path}
\title{returns content path}
\usage{
f_content_get_path(folder = "vignettes")
}
\description{
returns content path
}
\examples{
\dontrun{
 f_content_get_path()
}
}
