% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_content.R
\name{f_content_render}
\alias{f_content_render}
\title{renders content}
\usage{
f_content_render(overwrite = F, folder = "vignettes")
}
\arguments{
\item{overwrite}{logical, Default: F}
}
\examples{
\dontrun{
 f_content_render()
}
}
