% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_datatable.R
\name{f_datatable_minimal}
\alias{f_datatable_minimal}
\title{convert dataframe to DT:datatable with minimal features}
\usage{
f_datatable_minimal(df, escape = F)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{escape}{boolean, escape html content, Default:FALSE}
}
\value{
DT::datatable
}
\examples{
f_datatable_minimal(mtcars)
}
\seealso{
\code{\link[DT]{datatable}}
}
