% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_datatable.R
\name{f_datatable_universal}
\alias{f_datatable_universal}
\title{convert dataframe to DT:datatable inlcuding most usefull extensions
  and options}
\usage{
f_datatable_universal(df, round_perc = 1, round_sign = 2,
  count_cols_as_int = T, round_other_nums = NULL, page_length = 10)
}
\arguments{
\item{df}{dataframe}

\item{round_perc}{digits for percentages, Default: 1}

\item{round_sign}{digits for p_values, Default: 2}

\item{count_cols_as_int}{detect count columns, Default: T}

\item{round_other_nums}{round other numerical(double) columns to that digit. Will not
do anything if NULL, Default: NULL}

\item{page_length}{integer Default page length of table, Default: 10}
}
\value{
DT:datatable
}
\description{
inludes the features for excel and clipboard export, hide and
  unhide (column visibility), reorder columns per drag and drop, navigate
  table with arrow keys and prefix/suffix directed rounding of numerical
  values.
  \itemize{
  \item count_/ _count will round to 0
  \item p_val will round to Default: 2
  \item perc_/ _perc will round to Default: 1
  }
}
\examples{
data_ls = f_clean_data(mtcars)
f_stat_group_counts_percentages(data_ls, 'cyl') \%>\%
  f_datatable_universal()

f_stat_group_mean_medians(data_ls, 'cyl') \%>\%
  f_datatable_universal(round_other_nums = 2)
}
\seealso{
\code{\link[stringr]{str_detect}}
  \code{\link[DT]{datatable}},\code{\link[DT]{formatPercentage}},\code{\link[DT]{formatSignif}},\code{\link[DT]{formatRound}}
}
