% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_plot.R
\name{f_get_col_vector60}
\alias{f_get_col_vector60}
\title{FUNCTION_TITLE}
\usage{
f_get_col_vector60()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
