% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_get_col_vector60}
\alias{f_get_col_vector60}
\title{generate a most distinctive color scale}
\usage{
f_get_col_vector60()
}
\value{
vector with RGB colours
}
\description{
based on RColorBrewer colours of length 74
}
