% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_html.R
\name{f_html_breaks}
\alias{f_html_breaks}
\title{create a taglist with n lines of html line breaks}
\usage{
f_html_breaks(n)
}
\arguments{
\item{n}{number of line breaks}
}
\value{
taglist
}
\examples{
f_html_breaks(5)
}
\seealso{
\code{\link[htmltools]{br}},\code{\link[htmltools]{tagList}}
}
