% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_html.R
\name{f_html_get_title_from_Rmd}
\alias{f_html_get_title_from_Rmd}
\title{get title from Rmd file}
\usage{
f_html_get_title_from_Rmd(file_path)
}
\arguments{
\item{file_path}{path to Rmd file}
}
\value{
character vector
}
\description{
usefull if we want to provide additional iformation if we wat to
  link to an html file that was generates from a Rmd file.
}
\seealso{
\code{\link[readr]{read_lines}}
}
