% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_html.R
\name{f_html_link}
\alias{f_html_link}
\title{file path to html link}
\usage{
f_html_link(file_path, text)
}
\arguments{
\item{file_path}{file path}

\item{text}{link text}
}
\value{
character vector
}
\examples{
f_html_link(getwd(),'Working Directory')
}
