% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_html.R
\name{f_html_padding}
\alias{f_html_padding}
\title{add some padding around html objects}
\usage{
f_html_padding(obj, pad_before = 0, title = NULL, subtitle = NULL,
  caption = NULL, pad_after = 0, .f_htitle = htmltools::h3)
}
\arguments{
\item{obj}{html object such as DT:datatable() or plotly::ggplotly()}

\item{pad_before}{integer, Default: 0}

\item{title}{character vector, Default: ''}

\item{subtitle}{character vector, Default: ''}

\item{caption}{character vector, Default: ''}

\item{pad_after}{character vector, Default: 0}

\item{.f_htitle}{function, Default: htmltools::h3}
}
\value{
taglist
}
\examples{
f_html_padding(DT::datatable(mtcars),5,'mtcars Data','subtitle', 'caption', 8 )
}
\seealso{
\code{\link[htmltools]{tagList}},\code{\link[htmltools]{h3}},\code{\link[htmltools]{h4}},\code{\link[htmltools]{h6}}
}
