% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_html.R
\name{f_html_table_html_and_rmd_link}
\alias{f_html_table_html_and_rmd_link}
\title{create a DT::datatable that pairs Rmd and rendered html files.}
\usage{
f_html_table_html_and_rmd_link(wd = getwd(), folder_Rmd = getwd(),
  folder_html = getwd())
}
\arguments{
\item{wd}{working directory, Default: getwd()}

\item{folder_Rmd}{path to folder with Rmd files relative to wd (use '.' for
wd and '..' for parent directory of wd ), Default: getwd()}

\item{folder_html}{path to folder with html files relative to wd (use '.' for
wd and '..' for parent directory of wd ), Default: getwd()}
}
\value{
DT::datatable
}
\description{
Extracts the title from the Rmd file and generates one column
  with links to the Rmd and one column with links to the html files.
}
\examples{

wd = file.path( system.file(package = 'oetteR') )
folder_Rmd = 'vignettes'
folder_html = 'vignettes'

f_html_table_html_and_rmd_link( wd, folder_Rmd, folder_html )

}
