% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_append_2_list}
\alias{f_manip_append_2_list}
\title{append object to list}
\usage{
f_manip_append_2_list(l, x)
}
\arguments{
\item{l}{list}

\item{x}{object}
}
\value{
list
}
\description{
convenience function to replace  l[[length(l)+1]] = x
}
\examples{
l = list('a', 'b')
l = f_manip_append_2_list(l, 'c')
str(l)
}
