% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_bring_to_pos_range}
\alias{f_manip_bring_to_pos_range}
\title{bring vector to positice range}
\usage{
f_manip_bring_to_pos_range(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
vector
}
\description{
if min < 0, add abs(min) to all values
}
\examples{
vec = c( -2,0,2,4,6)
vec = f_manip_bring_to_pos_range( vec )
vec
}
