% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_double_2_int}
\alias{f_manip_double_2_int}
\title{converts columns of type double to integer if maximum number of
  decimal digits is zero}
\usage{
f_manip_double_2_int(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
tibble
}
\examples{
as_tibble(mtcars)
f_manip_double_2_int(mtcars)
}
