% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_get_most_common_level}
\alias{f_manip_get_most_common_level}
\title{get most common level from vector}
\usage{
f_manip_get_most_common_level(x)
}
\arguments{
\item{x}{factor vector}
}
\value{
character vector
}
\seealso{
\code{\link[broom]{tidy}}
}
