% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_get_response_variable_from_formula}
\alias{f_manip_get_response_variable_from_formula}
\title{get response variable from formula}
\usage{
f_manip_get_response_variable_from_formula(formula)
}
\arguments{
\item{formula}{formula}
}
\value{
character vector
}
\description{
get response variable from formula
}
\seealso{
\code{\link{f_manip_get_variables_from_formula}}
}
