% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_get_variables_from_formula}
\alias{f_manip_get_variables_from_formula}
\title{get variables from formula}
\usage{
f_manip_get_variables_from_formula(formula)
}
\arguments{
\item{formula}{formula}
}
\value{
character vector
}
\description{
get variables from formula
}
\examples{
f = foo~bar1 + bar2

vars = f_manip_get_variables_from_formula(f)
response_var = f_manip_get_response_variable_from_formula(f)

}
\seealso{
\code{\link{f_manip_get_response_variable_from_formula}}
}
