% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_matrix_2_tibble}
\alias{f_manip_matrix_2_tibble}
\title{converts matrices to tibble, preserving row.names}
\usage{
f_manip_matrix_2_tibble(x)
}
\arguments{
\item{x}{any variable}
}
\value{
a tibble or if the input variable is neither matrix dataframe or
  tibble the original input object.
}
\description{
row.names are added as row_names column as the first column of
  the tibble. Function does not fail when object cannot be converted to
  tibble thus can be used to map over lists with various variable types such
  as modells and objects.
}
\examples{

mat = as.matrix(mtcars)
head( mat, 10)
f_manip_matrix_2_tibble( mat )

# convert all matrices from a list
pca = prcomp( mtcars ) \%>\%
 map( f_manip_matrix_2_tibble )
pca

}
