% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_summarize_2_median_and_most_common_factor}
\alias{f_manip_summarize_2_median_and_most_common_factor}
\title{takes a data_ls list created by f_clean_data() and returns a list with
  all medians for numerical and most common level for categorical variables.}
\usage{
f_manip_summarize_2_median_and_most_common_factor(data_ls)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}
}
\value{
list\describe{
  \item({data}{summarized data as dataframe})
  \item({data_boxcox}{summarized boxcox data as dataframe})
}
}
\examples{
summarized_ls = f_clean_data(mtcars) \%>\%
  f_boxcox() \%>\%
  f_manip_summarize_2_median_and_most_common_factor()

summarized_ls$data
summarized_ls$boxcox_data
}
