% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_manip.R
\name{f_manip_transpose_tibble}
\alias{f_manip_transpose_tibble}
\title{transpose a tibble}
\usage{
f_manip_transpose_tibble(tib)
}
\arguments{
\item{tib}{tibble}
}
\value{
tibble
}
\description{
transpose a tibble, values in first column will become column
 titles. Row names will be converted to first columns
}
\examples{

tib = mtcars \%>\%
  as_tibble() \%>\%
  f_manip_transpose_tibble()
tib
}
