% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_var_dep.R,
%   R/f_model_vis_variable_dependency.R
\name{f_model_add_predictions_2_grid_regression}
\alias{f_model_add_predictions_2_grid_regression}
\alias{f_model_add_predictions_2_grid_regression}
\title{add predictions to grid (regression models)}
\usage{
f_model_add_predictions_2_grid_regression(grid, m, var)

f_model_add_predictions_2_grid_regression(grid, m, var)
}
\arguments{
\item{grid}{grid containing all variables used for the model}

\item{m}{model}

\item{var}{character vector denoting response variable}

\item{grid}{grid containing all variables used for the model}

\item{m}{model}

\item{var}{character vector denoting response variable}
}
\value{
grid

grid
}
\description{
wrapper for modelr::add_predictions

wrapper for modelr::add_predictions
}
\examples{
data_ls = f_clean_data(mtcars)
formula = disp~hp+mpg
m = lm(formula, data_ls$data)
grid = f_model_data_grid(data_ls, formula, 'hp', 10) \%>\%
  f_model_add_predictions_2_grid_regression( m, 'disp')
data_ls = f_clean_data(mtcars)
formula = disp~hp+mpg
m = lm(formula, data_ls$data)
grid = f_model_data_grid(data_ls, formula, 'hp', 10) \%>\%
  f_model_add_predictions_2_grid_regression( m, 'disp')
}
\seealso{
\code{\link[modelr]{add_predictions}}

\code{\link[modelr]{add_predictions}}
}
