% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_var_dep.R
\name{f_model_data_grid}
\alias{f_model_data_grid}
\title{generates a data grid based on a formula}
\usage{
f_model_data_grid(col_var, data_ls, formula, n = 500)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{formula}{formula}
}
\value{
dataframe
}
\description{
the range of one specified variable is expanded, while all other
  variables are set to the most common values. Similar to modelr::data_grid
  but it can deal with factors.
}
\examples{
data_ls = f_clean_data(mtcars)
formula = disp~cyl+mpg+hp
f_model_data_grid( 'mpg', data_ls, formula,  10 )
}
