% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_imp.R
\name{f_model_importance}
\alias{f_model_importance}
\title{model importance}
\usage{
f_model_importance(m, data)
}
\arguments{
\item{m}{model}

\item{data}{training data}
}
\value{
tibble
}
\description{
supports rpart, randomForest, svm, will return NULL for other models
}
\examples{
pl = pipelearner::pipelearner(mtcars) \%>\%
  pipelearner::learn_models( twidlr::rpart, disp~. ) \%>\%
  pipelearner::learn_models( twidlr::randomForest, disp~. ) \%>\%
  pipelearner::learn_models( twidlr::svm, disp~. ) \%>\%
  pipelearner::learn() \%>\%
  mutate( imp = map2(fit, train, f_model_importance) )
pl$imp
}
