% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats_importance.R
\name{f_model_importance_pl_add_plots_regression}
\alias{f_model_importance_pl_add_plots_regression}
\title{add plots based on variable importance to pipelearner dataframe}
\usage{
f_model_importance_pl_add_plots_regression(pl, data, m, ranked_variables,
  response_var, title, variable_color_code, formula, data_ls,
  var_dep_limit = 10, var_dep_log_y = F, tabplot_limit = 12)
}
\arguments{
\item{pl}{a dataframe containing the columns for data, m, ranked_variables, response_var and title}

\item{data}{symbol (unquoted name) of data column in pl}

\item{m}{symbol (unquoted name) of data column in pl}

\item{ranked_variables}{symbol (unquoted name) of data column in pl}

\item{response_var}{symbol (unquoted name) of data column in pl}

\item{title}{symbol (unquoted name) of data column in pl}

\item{variable_color_code}{dataframe created by f_plot_color_code_variables()}

\item{formula}{fomula that was used to construct model}

\item{data_ls}{data_ls list object wontaining the whole of the original data}

\item{var_dep_limit}{number of variables to be plotted on dependency plot}

\item{var_dep_log_y}{should y axis of dependency plot be logarithmic}

\item{tabplot_limit}{number of variables to be plotted on tabplot}
}
\value{
dataframe
}
\description{
adds a bar plot of the ranked variables, a tabplot sorted by the
  target variable and a dependency plot (response variable vs the sequential
  range of one of the predictor variables while all other predictors are kept
  constant at mean values).
}
\examples{

data_ls = f_clean_data(mtcars)
form = disp~cyl+mpg+hp
variable_color_code = f_plot_color_code_variables(data_ls)

pl = pipelearner::pipelearner(data_ls$data) \%>\%
  pipelearner::learn_models( twidlr::rpart, form ) \%>\%
  pipelearner::learn_models( twidlr::randomForest, form ) \%>\%
  pipelearner::learn_models( twidlr::svm, form ) \%>\%
  pipelearner::learn() \%>\%
  mutate( imp = map2(fit, train, f_model_importance)
          , title = paste(model, models.id, train_p) ) \%>\%
  f_model_importance_pl_add_plots_regression(  data                  = train
                                               , m                   = fit
                                               , ranked_variables    = imp
                                               , title               = title
                                               , response_var        = target
                                               , variable_color_code = variable_color_code
                                               , formula             = form
                                               , data_ls             = data_ls
                                               , var_dep_limit       = 10
                                               , var_dep_log_y       = T
                                               , tabplot_limit       = 12 )

}
\seealso{
\code{\link{f_model_importance_plot}}
\code{\link{f_model_importance_plot_tableplot}}
\code{\link{f_model_plot_variable_dependency_regression}}
}
