% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_imp.R
\name{f_model_importance_pl_plots_as_html}
\alias{f_model_importance_pl_plots_as_html}
\title{print plots of variable importance in modelling dataframe to html}
\usage{
f_model_importance_pl_plots_as_html(pl, prefix = NULL, quiet = FALSE)
}
\arguments{
\item{pl}{modelling dataframe containing the following columns 'imp_plot', 'imp_plot_dep', 'imp_tabplot', 'title'}

\item{prefix}{character vector file name prefix for html files, Default: NULL}

\item{quiet}{boolean, suppresses output to console by render function, Default: FALSE}
}
\value{
html files in working directory
}
\description{
should execute f_model_importance_pl_add_plots_regression() on
  modelling dataframe first
}
\examples{
\dontrun{
    data_ls = f_clean_data(mtcars)
form = disp~cyl+mpg+hp
variable_color_code = f_plot_color_code_variables(data_ls)

pl = pipelearner::pipelearner(data_ls$data) \%>\%
 pipelearner::learn_models( twidlr::rpart, form ) \%>\%
 pipelearner::learn_models( twidlr::randomForest, form ) \%>\%
 pipelearner::learn_models( twidlr::svm, form ) \%>\%
 pipelearner::learn() \%>\%
 mutate( imp = map2(fit, train, f_model_importance)
         , title = paste(model, models.id, train_p) ) \%>\%
 f_model_importance_pl_add_plots_regression(  data                  = train
                                              , m                   = fit
                                              , ranked_variables    = imp
                                              , title               = title
                                              , response_var        = target
                                              , variable_color_code = variable_color_code
                                              , formula             = form
                                              , data_ls             = data_ls
                                              , var_dep_limit       = 10
                                              , var_dep_log_y       = T
                                              , tabplot_limit       = 12) \%>\%
 f_model_importance_pl_plots_as_html( prefix = 'test_oetteR_html_')

files = dir() \%>\%
 .[ startsWith(., 'test_oetteR_html_') ]

file.remove( files )


}
}
\seealso{
\code{\link[htmltools]{tagList}}
}
