% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_imp.R
\name{f_model_importance_randomForest}
\alias{f_model_importance_randomForest}
\title{extract variable importance for randomForest model}
\usage{
f_model_importance_randomForest(m)
}
\arguments{
\item{m}{model of class randomForest}
}
\value{
dataframe
}
\description{
extract variable importance for randomForest model
}
\examples{
#regression
m = randomForest::randomForest(mtcars, disp~.)
f_model_importance_randomForest(m)

#classification
data_ls = f_clean_data(mtcars)
m = randomForest::randomForest(data_ls$data, cyl~.)
f_model_importance_randomForest(m)

}
