% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats.R
\name{f_model_importance_rpart}
\alias{f_model_importance_rpart}
\title{extract variable importance for rpart}
\usage{
f_model_importance_rpart(m, ...)
}
\arguments{
\item{m}{model of class rpart}
}
\value{
dataframe
}
\examples{

#regression
m = twidlr::rpart(mtcars, disp~.)
f_model_importance_rpart(m)

#classification
data_ls = f_clean_data(mtcars)
m = twidlr::rpart(data_ls$data, cyl~.)
f_model_importance_rpart(m)

}
