% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats.R
\name{f_model_importance_svm}
\alias{f_model_importance_svm}
\title{extract variable importance for svm}
\usage{
f_model_importance_svm(m)
}
\arguments{
\item{m}{model of class svm}
}
\value{
dataframe
}
\details{
uses 1D-SA 1 dimensional sensitivity analysis using
  rminer::Importance()
}
\examples{

#regression
m = twidlr::svm(mtcars, disp~.)
f_model_importance_svm(m)

#classification
data_ls = f_clean_data(mtcars)
m = twidlr::svm(data_ls$data, cyl~.)
f_model_importance_svm(m)

}
