% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_var_dep.R
\name{f_model_seq_range}
\alias{f_model_seq_range}
\title{generates sequence of variable spanning from min to max}
\usage{
f_model_seq_range(data_ls, col_var, n = 500)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_var}{character vector denoting variable}

\item{n}{integer number of intermediate data points, Default: 500}
}
\value{
vector
}
\description{
similar to modelr::seq_range but can handle categorical variables
}
\examples{
data_ls = f_clean_data(mtcars)
col_var = 'disp'
f_model_seq_range( data_ls, col_var, 10)
}
