% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_pca.R
\name{f_pca}
\alias{f_pca}
\title{calculate principle components for a dataset}
\usage{
f_pca(data_ls, center = T, scale = T, use_boxcox_tansformed_vars = T,
  include_ordered_categoricals = T, threshold_vae_for_pc_perc = 2.5)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{center}{boolean, Default: T}

\item{scale}{boolean, Default: T}

\item{use_boxcox_tansformed_vars}{boolean, Default: T}
}
\value{
a list with the original data complemented with the principle
 component vector data of each observation and an object returned by prcomp()
 supplemented with some extra features \item{data}{dataframe} \item{pca}{pca
 object created by prcomp()}

 \strong{added features of pca}: \item{cos2}{The squared rotation vectors.A
 value between 0 and 1 denotes the amount of contribution of a variable to a
 specific principle component} \item{vae}{percent variance explained}
 \item{contrib_abs_perc}{The absolute contribution of one variable to the
 variance explained by one principle component in percent. The total
 contibution adds up to the total contibution of the principle componaent in
 percent. } \item{contrib_abs_perc_reduced}{as above but variables
 contibuting less than 2.5 percent are grouped}
 \item{threshold_vae_for_pc_perc}{principle components that explain less
 percent variance than this threshold are dropped}
}
\description{
This function is an extended wrapper for prcomp(). I takes a
 data_ls object created by f_clean_data and calculates the contribution of
 each variable to each principle component in percent.
}
\details{
\href{http://www.sthda.com/english/wiki/principal-component-analysis-how-to-reveal-the-most-important-variables-in-your-data-r-software-and-data-mining}{Blog
post explaining how to calculate contributions}
}
\examples{
pca_ls = f_clean_data(mtcars) \%>\%
f_boxcox() \%>\%
f_pca()

}
\seealso{
\code{\link[stats]{prcomp}}
}
