% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_pca.R
\name{f_pca_plot_variance_explained}
\alias{f_pca_plot_variance_explained}
\title{plot varaince explained of principle components}
\usage{
f_pca_plot_variance_explained(pca_ls, threshold_vae_for_pc_perc = 2.5)
}
\arguments{
\item{pca_ls}{list created by f_pca()}
}
\value{
plotly graph
}
\examples{
p = f_clean_data(mtcars) \%>\%
  f_boxcox() \%>\%
  f_pca() \%>\%
  f_pca_plot_variance_explained()
p
}
