% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R, R/f_plot_alluvial.R
\name{f_plot_alluvial}
\alias{f_plot_alluvial}
\alias{f_plot_alluvial_1v1}
\title{plot alluvial}
\usage{
f_plot_alluvial(data, variables = names(data), max_variables = 20,
  bins = 5, bin_labels = c("LL", "ML", "M", "MH", "HH"),
  order_levels = NULL, fill_by = "first_variable",
  col_vector = RColorBrewer::brewer.pal(name = "Dark2", n = 8))

f_plot_alluvial_1v1(data, col_x, col_y, col_id, col_fill = NULL,
  max_variables = 20, bins = 5, bin_labels = c("LL", "ML", "M", "MH",
  "HH"), order_levels_y = NULL, order_levels_x = NULL,
  order_levels_fill = NULL, complete = TRUE, fill_by = "first_variable",
  col_vector = RColorBrewer::brewer.pal(name = "Dark2", n = 8))
}
\arguments{
\item{data}{a dataframe}

\item{variables}{vector denoting names and order of the plotted variables,
Default: names(data)}

\item{max_variables}{maximum number of variables, Default: 20}

\item{bins}{number of bins for numerical variables, Default: 5}

\item{bin_labels}{labels for the bins from low to high, Default: c("LL",
"ML", "M", "MH", "HH")}

\item{order_levels}{character vector denoting levels to be reorderer from low to high}

\item{fill_by}{one_of(c('first_variable', 'last_variable', 'all_flows',
'values')), Default: 'first_variable'}

\item{col_vector}{vector with HEX color codes, Default:
RColorBrewer::brewer.pal(name = "Dark2", n = 8)}

\item{data}{a dataframe}

\item{variables}{vector denoting names and order of the plotted variables,
Default: names(data)}

\item{max_variables}{maximum number of variables, Default: 20}

\item{bins}{number of bins for numerical variables, Default: 5}

\item{bin_labels}{labels for the bins from low to high, Default: c("LL",
"ML", "M", "MH", "HH")}

\item{order_levels}{character vector denoting levels to be reorderer from low to high}

\item{fill_by}{one_of(c('first_variable', 'last_variable', 'all_flows',
'values')), Default: 'first_variable'}

\item{col_vector}{vector with HEX color codes, Default:
RColorBrewer::brewer.pal(name = "Dark2", n = 8)}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
plots a dataframe as an alluvial plot. All numerical variables
  are scaled, centered and YeoJohnson transformed before binning.

plots a dataframe as an alluvial plot. All numerical variables
  are scaled, centered and YeoJohnson transformed before binning.
}
\details{
DETAILS

DETAILS
}
\seealso{
\code{\link[RColorBrewer]{brewer.pal}}
  \code{\link[forcats]{fct_relevel}}
  \code{\link[ggalluvial]{geom_flow}},\code{\link[ggalluvial]{geom_stratum}}

\code{\link[RColorBrewer]{brewer.pal}}
  \code{\link[forcats]{fct_relevel}}
  \code{\link[ggalluvial]{geom_flow}},\code{\link[ggalluvial]{geom_stratum}}
}
