% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_col_vector74}
\alias{f_plot_col_vector74}
\title{generate a most distinctive color scale}
\usage{
f_plot_col_vector74()
}
\value{
vector with HEX colours
}
\description{
based on RColorBrewer colours of length 74
}
