% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_color_code_variables}
\alias{f_plot_color_code_variables}
\title{color code all variables in a data_ls list.}
\usage{
f_plot_color_code_variables(data_ls, col_vector = f_plot_col_vector74())
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_vector}{character vector denoting Hexcode colors, Default: f_plot_col_vector74()}
}
\value{
tibble
\itemize{
  \item{variable}{}
  \item{color}{HEX code color}
  }
}
\description{
color coding is stable the same data_ls list gets the same
  coding with every function call. Assigns the same colors to the boxcox
  transformed and untransformed variant of a variable.
}
\examples{
f_clean_data(mtcars) \%>\%
  f_boxcox() \%>\%
  f_plot_color_code_variables() \%>\%
  print()
}
\seealso{
\code{\link[stringr]{str_replace_all}}
}
