% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_generate_comparison_pairs}
\alias{f_plot_generate_comparison_pairs}
\title{generates comparison pairs for `ggpubr::stat_compare_means()`}
\usage{
f_plot_generate_comparison_pairs(data, col_var, col_group, thresh = 0.05)
}
\arguments{
\item{data}{dataframe}

\item{col_var}{character vector denoting variable column}

\item{col_group}{character vector denoting grouping column}

\item{thresh}{double, Default: 0.05}
}
\value{
list
}
\description{
generates all possible pairs and filters according to wilcox p_value
}
\examples{

f_plot_generate_comparison_pairs( mtcars, 'disp', 'cyl' )

}
\seealso{
\code{\link[stringr]{str_split}}
 \code{\link[rlang]{UQ}}
}
