% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_plot.R
\name{f_plot_hist}
\alias{f_plot_hist}
\title{FUNCTION_TITLE}
\usage{
f_plot_hist(variable, data_ls, group = "None", graph_type = c("violin",
  "bar", "line"), y_axis = c("count", "density"), auto_range = T,
  n_breaks = 30, x_min = 0, x_max = 100)
}
\arguments{
\item{variable}{PARAM_DESCRIPTION}

\item{data_ls}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{graph_type}{PARAM_DESCRIPTION, Default: c("violin", "bar", "line")}

\item{y_axis}{PARAM_DESCRIPTION, Default: c("count", "density")}

\item{auto_range}{PARAM_DESCRIPTION, Default: T}

\item{n_breaks}{PARAM_DESCRIPTION, Default: 30}

\item{x_min}{PARAM_DESCRIPTION, Default: 0}

\item{x_max}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{

#plot single variable
data_ls = f_clean_data(mtcars)
f_plot_hist('disp', data_ls)

}
\seealso{
\code{\link[stringr]{str_extract_all}},\code{\link[stringr]{str_c}}
}
