% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_obj_2_html}
\alias{f_plot_obj_2_html}
\title{generate a separate html file from a shiny taglist}
\usage{
f_plot_obj_2_html(obj_list, type, output_file, title = "Plots", ...)
}
\arguments{
\item{obj_list}{htmltools::tagList}

\item{output_file}{file_name of the html file, without .html suffix}

\item{title}{character vector of html document title, Default: 'Plots'}
}
\description{
html objects generated by plotly or DT can be stored in
  htmltools::tagList. It can be printed in a html file to display the objects.
}
\examples{
#returns a htmltools::taglist with DT::datatables and plotly plots
taglist = f_clean_data(mtcars) \%>\%
  f_boxcox() \%>\%
  f_pca() \%>\%
  f_pca_plot_components()

f_plot_obj_2_html(taglist, type = "taglist", output_file =  'test_me', title = 'Plots')
file.remove('test_me.html')
}
