% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_pretty_points}
\alias{f_plot_pretty_points}
\title{plot prettier dot plot}
\usage{
f_plot_pretty_points(df, col_x, col_y, col_facet = NULL, size = 4,
  title = NULL, x_title = col_x, y_title = col_y, ...)
}
\arguments{
\item{df}{datafram containing x,y pairs}

\item{col_x}{character vector denoting x axis values}

\item{col_y}{character vector denoting y axis values}

\item{col_facet}{character vector denoting facetting column}

\item{size}{size of points, Default: 4}

\item{...}{arguments passed to facet_wrap()}
}
\value{
plot
}
\description{
color is contineoulsy scaled based on PC1 values and alpha
  values depend on point density.
}
\details{
Code adapted from \url{https://drsimonj.svbtle.com/pretty-scatter-plots-with-ggplot2}
}
\examples{
df = ggplot2::diamonds \%>\%
  sample_n(2500)
col_x = 'carat'
col_y = 'price'
col_facet = 'cut'

f_plot_pretty_points(df, col_x, col_y, col_facet, title = 'price of diamonds by carat')
}
\seealso{
\code{\link[fields]{interp.surface}} \code{\link[MASS]{kde2d}}
}
