% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_profit.R
\name{f_plot_profit_bars_plus_area}
\alias{f_plot_profit_bars_plus_area}
\title{plot revenues cost and profit development over time with bars for
  revenue and costs and an area chart for profit.}
\usage{
f_plot_profit_bars_plus_area(data, col_revenue, col_cost, col_time,
  now = max(data[, col_time]), unit_time = "years",
  unit_value = "CHF", title = "", alpha_past = 1,
  alpha_future = 0.5, alpha_past_area = 0.9, alpha_future_area = 0.7)
}
\arguments{
\item{data}{datafram}

\item{col_revenue}{character vector denoting revenue column}

\item{col_cost}{character vector denoting cost column}

\item{col_time}{character vector denoting time column}

\item{now}{integer denoting a time which should be regarded as the
breakpoint, Default: max(data[, col_time])}

\item{unit_time}{character vector, will label y-axis, Default: 'years'}

\item{unit_value}{character vector, will label x-axis, Default: 'CHF'}

\item{title}{character vector, will be title label, Default: ''}

\item{alpha_past}{double between 0 and 1 will determine alpha value for fill
under the curve before the breakpoint, Default: 1}

\item{alpha_future}{double between 0 and 1 will determine alpha value for
fill under the curve after the breakpoint, Default: 0.5}

\item{alpha_past_area}{as alpha_past but for area only, Default: 0.9}

\item{alpha_future_area}{as alpha_future but for area only, Default: 0.7#'}
}
\value{
plot (to some extent plotly compatible)
}
\description{
the function can graphically devide the chart into two periods
  e.g. past and future.
}
\details{
to some extent plotly compatible
}
\examples{
data = tibble( time     = c(0,1,2,3,4,5,6,7,8,9,10,11,12)
              , revenue = - time^2 + time * 12
              , cost    = revenue * 0.4 * -1
             )
data[1,'cost'] = -10
data

print( f_plot_profit_bars_plus_area( data, 'revenue', 'cost', 'time') )
print( f_plot_profit_bars_plus_area( data, 'revenue', 'cost', 'time', now = 5) )

#clv figure for presenation
p = f_plot_profit_bars_plus_area( data, 'revenue', 'cost', 'time', now = 5, alpha_past_area = 0) +
  theme( panel.grid.major  = element_blank()
        , panel.grid.minor = element_blank()
        , axis.text        = element_blank()
        )+
  labs(x = '', y = '')
print(p)

}
