% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_plot_profit_lines}
\alias{f_plot_profit_lines}
\title{plot revenues cost and profit development over time as an area chart.}
\usage{
f_plot_profit_lines(data, col_revenue, col_cost, col_time, now = max(data[,
  col_time]), unit_time = "years", unit_value = "CHF", title = "",
  alpha_past = 1, alpha_future = 0.5)
}
\arguments{
\item{data}{datafram}

\item{col_revenue}{character vector denoting revenue column}

\item{col_cost}{character vector denoting cost column}

\item{col_time}{character vector denoting time column}

\item{now}{integer denoting a time which should be regarded as the
breakpoint, Default: max(data[, col_time])}

\item{unit_time}{character vector, will label y-axis, Default: 'years'}

\item{unit_value}{character vector, will label x-axis, Default: 'CHF'}

\item{title}{character vector, will be title label, Default: ''}

\item{alpha_past}{double between 0 and 1 will determine alpha value for fill
under the curve before the breakpoint, Default: 1}

\item{alpha_future}{double between 0 and 1 will determine alpha value for fill
under the curve after the breakpoint, Default: 0.5}
}
\value{
plot (is not plotly compatibel)
}
\description{
the function can graphically devide the chart into two periods
  e.g. past and future.
}
\details{
not plotly compatibel
}
\examples{
data = tibble( time     = c(0,1,2,3,4,5,6,7,8,9,10)
              , revenue = - time^2 + time * 12
              , cost    = revenue * 0.4 * -1
             )

print( f_plot_profit_lines( data, 'revenue', 'cost', 'time') )
print( f_plot_profit_lines( data, 'revenue', 'cost', 'time', now = 5) )

}
