% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_plot.R
\name{f_plot_time}
\alias{f_plot_time}
\title{FUNCTION_TITLE}
\usage{
f_plot_time(variable, time_variable, data_ls, .f = mean,
  time_variable_as_factor = F, group = NULL, normalize = F,
  time_unit = "day")
}
\arguments{
\item{variable}{PARAM_DESCRIPTION}

\item{time_variable}{PARAM_DESCRIPTION}

\item{data_ls}{PARAM_DESCRIPTION}

\item{.f}{PARAM_DESCRIPTION, Default: mean}

\item{time_variable_as_factor}{PARAM_DESCRIPTION, Default: F}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{normalize}{PARAM_DESCRIPTION, Default: F}

\item{time_unit}{PARAM_DESCRIPTION, Default: 'day'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
