% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_pl_regression}
\alias{f_predict_pl_regression}
\title{adds predictions to learned pipelearner dataframe}
\usage{
f_predict_pl_regression(pl, cols_id = NULL, formula = NULL,
  col_model = "fit", col_target = "target", data_test = "test",
  data_train = "train")
}
\arguments{
\item{pl}{learned pipelearner dataframe}

\item{cols_id}{character vector naming id columns}
}
\value{
dataframe
}
\examples{
form = as.formula( 'disp~cyl+mpg')

pl = mtcars \%>\%
  mutate(names = row.names(.)) \%>\%
  pipelearner::pipelearner() \%>\%
  pipelearner::learn_models( twidlr::rpart, form ) \%>\%
  pipelearner::learn_models( twidlr::randomForest, form ) \%>\%
  pipelearner::learn_models( twidlr::svm, form ) \%>\%
  pipelearner::learn_models( gamlss::gamlss, form ) \%>\%
  pipelearner::learn() \%>\%
  f_predict_pl_regression( cols_id = 'names' )
}
\seealso{
\code{\link{f_predict_regression_add_predictions}}
}
