% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_pl_regression_summarize}
\alias{f_predict_pl_regression_summarize}
\title{summarize prediction by f_predict_pl_regression()}
\usage{
f_predict_pl_regression_summarize(pl)
}
\arguments{
\item{pl}{pipelearner dataframe with nested predictions}
}
\value{
dataframe with mape, mea, rtmse and median versions
}
\description{
use this function to get a quick summary of pipelearner
  dataframe with unnested predictions. Will group by title
}
\examples{
form = as.formula( 'disp~cyl+mpg')

pl = mtcars \%>\%
  mutate(names = row.names(.)) \%>\%
  pipelearner::pipelearner() \%>\%
  pipelearner::learn_models( rpart::rpart, form ) \%>\%
  pipelearner::learn_models( randomForest::randomForest, form ) \%>\%
  pipelearner::learn_models( e1071::svm, form ) \%>\%
  pipelearner::learn() \%>\%
  f_predict_pl_regression( 'names' ) \%>\%
  unnest( preds , .drop = FALSE ) \%>\%
  mutate( title = model ) \%>\%
  f_predict_pl_regression_summarize()

pl

}
\seealso{
\code{\link{f_predict_pl_regression}}
}
