% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_plot_regression_alluvials}
\alias{f_predict_plot_regression_alluvials}
\title{plot residuals of different models as aaluvials}
\usage{
f_predict_plot_regression_alluvials(data, col_id, col_title = "title",
  col_pred = "pred", col_obs = "target1", ...)
}
\arguments{
\item{data}{dataframe}

\item{col_id}{character vecotr dentoing id column}

\item{col_title}{character vector denoting model title column, Default:
'title'}

\item{col_pred}{character vector denoting prediciont column, Default: 'pred'}

\item{col_obs}{character vector denoting column with observed values,
Default: 'target1'}

\item{...}{additional arguments passed to f_plot_alluvial_1v1}
}
\value{
plot
}
\description{
Residuals will be  binned using boxplotstats with the
  modification, that the median will be set to zero. If quantile boundaries
  do not fall into sensible ranges they will be replaced by the standard
  error (SE). For example if we have the following boxplotstats -2, 1, 3, 5,
  8 with an SE of 0.5 we will modify the boundaries as following. -2, -0.5,
  0, 3, 5 . The reason is that we want to be able to follow observations with
  positive or negative residuals through the alluvial plot in order to judge
  emerging patterns. The models will be sorted by MSE and the alluvial plot
  will be flipped with the model with the lowest MSE on top.
}
\examples{

  form = as.formula( 'displacement~cylinders+mpg')

df = ISLR::Auto \%>\%
  mutate( name = paste( name, row_number() ) ) \%>\%
  pipelearner::pipelearner() \%>\%
  pipelearner::learn_models( rpart::rpart, form ) \%>\%
  pipelearner::learn_models( randomForest::randomForest, form ) \%>\%
  pipelearner::learn_models( e1071::svm, form ) \%>\%
  pipelearner::learn() \%>\%
  f_predict_pl_regression( 'name' ) \%>\%
  unnest(preds)

f_predict_plot_regression_alluvials(df
                                    , col_id = 'name'
                                    , col_title = 'model'
                                    , col_pred = 'pred'
                                    , col_obs = 'target1')

}
\seealso{
\code{\link[RColorBrewer]{brewer.pal}}
}
