% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_plot_regression_distribution}
\alias{f_predict_plot_regression_distribution}
\title{Plot distribution of model predictions vs observed}
\usage{
f_predict_plot_regression_distribution(data, col_title = "title",
  col_pred = "pred", col_obs = "target1", bins = 60, ...)
}
\arguments{
\item{data}{dataframE}

\item{col_title}{character vector denoting title column, Default: 'title'}

\item{col_pred}{character vector denoting column with predictions, Default:
'preds'}

\item{col_obs}{character vecor denoting column with observed values, Default:
'target1'}

\item{bins}{number of bins used for histograms, Default: 60}

\item{...}{additional arguments passed to the facet_wrap function of the
histogramss}
}
\value{
list with two plots
}
\description{
takes a dataframe with predictions and a title column and
  returns a list with one violin and one histogram plot to compare
  distributions.
}
\examples{

  form = as.formula( 'displacement~cylinders+mpg')

  df = ISLR::Auto \%>\%
    pipelearner::pipelearner() \%>\%
    pipelearner::learn_models( rpart::rpart, form ) \%>\%
    pipelearner::learn_models( randomForest::randomForest, form ) \%>\%
    pipelearner::learn_models( e1071::svm, form ) \%>\%
    pipelearner::learn() \%>\%
    f_predict_pl_regression( 'name' ) \%>\%
    unnest(preds)

  f_predict_plot_regression_distribution(df
                                         , col_title = 'model'
                                         , col_pred = 'pred'
                                         , col_obs = 'target1')

}
\seealso{
\code{\link[rlang]{UQ}} \code{\link[RColorBrewer]{brewer.pal}}
}
