% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_regression_add_predictions}
\alias{f_predict_regression_add_predictions}
\title{adds predictions, residuals, abolute residuals, squared residuals and
  absolute percent error to a dataframe}
\usage{
f_predict_regression_add_predictions(df, m, col_target, cols_id)
}
\arguments{
\item{df}{dataframe}

\item{m}{regression model}

\item{col_target}{character vector naming targtet/response variable}

\item{cols_id}{character vector naming id columns}
}
\value{
dataframe
}
\description{
absolute percent error = (abs(resid/pred)*100 )
}
\examples{
df = mtcars \%>\%
mutate(names = row.names(.))
m = rpart::rpart(disp~., df)
f_predict_regression_add_predictions(df, m, 'disp', 'names')
}
\seealso{
\code{\link[modelr]{add_predictions}},\code{\link[modelr]{add_residuals}}
}
