% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_shiny.R
\name{f_shiny_multiview}
\alias{f_shiny_multiview}
\title{run multiview shiny app}
\usage{
f_shiny_multiview(data = NULL)
}
\arguments{
\item{data}{dataframe, will let you choose between a variety of sample data
sets if NULL, Default: NULL}
}
\description{
shiny tool to quickly analyze diffferences between groups in
  large datasets
}
\seealso{
\code{\link[rmarkdown]{run}}
}
