% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_shiny.R
\name{f_shiny_som}
\alias{f_shiny_som}
\title{self organizing map shiny app}
\usage{
f_shiny_som(data = NULL)
}
\arguments{
\item{data}{dataframe, will let you choose between a variety of sample data
sets if NULL, Default: NULL}
}
\description{
shiny tool to build self organizing maps with subsequent
  clustering restricted to adjacent map segements. See http://rpubs.com/erblast/SOM.
}
\seealso{
\code{\link[rmarkdown]{run}}
}
