% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sim.R
\name{f_sim_profit}
\alias{f_sim_profit}
\title{simulate profit}
\usage{
f_sim_profit(output_file = "profit_simulation", path = ".",
  params = "ask")
}
\arguments{
\item{output_file}{PARAM_DESCRIPTION, Default: 'profit_simulation'}

\item{path}{PARAM_DESCRIPTION, Default: '.'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
using the following parameters:
 \itemize{
 \item retention rate
 \item new customers acquired per year
 \item present number of customers
 \item fixed cost
 \item profit per customer per year
 }
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 f_sim_profit( path = tempdir() )
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
