% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_chi_square}
\alias{f_stat_chi_square}
\title{generate a datatframe with chi square results from a data_ls list}
\usage{
f_stat_chi_square(data_ls, col_group)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_group}{character vector denoting grouping variable}
}
\value{
dataframe
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
data_ls = f_clean_data(mtcars)
df_chi_squ = f_stat_chi_square(data_ls, 'cyl')
df_chi_squ
}
\seealso{
\code{\link[purrr]{is_empty}},\code{\link[purrr]{map}},\code{\link[purrr]{map_dbl}}
}
