% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_combine_anova_with_chi_square}
\alias{f_stat_combine_anova_with_chi_square}
\title{combines anova with chi square results into single dataframe}
\usage{
f_stat_combine_anova_with_chi_square(df_anova = NULL,
  df_chi_square = NULL)
}
\arguments{
\item{df_anova}{dataframe created with f_stat_anova, Default: NULL}

\item{df_chi_square}{dataframe created with f_stat_chi_square(), Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
keeps wither the anova or the kruskal p value depending on the
  results of the shapiro test (shapiro_stat > 0.9 p_val > 0.05) and keeps
  the difference of percent of the mean. Still works if one of the input
  dataframes is NULL
}
\details{
DETAILS
}
\examples{
 data_ls = f_clean_data(mtcars)
 df_chi_squ = f_stat_chi_square(data_ls, 'cyl')
 df_anova = f_stat_anova(data_ls, 'cyl')
 df_comb = f_stat_combine_anova_with_chi_square(df_anova, df_chi_squ)
 df_comb
 df_comb = f_stat_combine_anova_with_chi_square(df_anova)
 df_comb
 df_comb = f_stat_combine_anova_with_chi_square(df_chi_square = df_chi_squ)
 df_comb
}
