% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_diff_of_means_medians}
\alias{f_stat_diff_of_means_medians}
\title{calculates maximum difference in group means and medians}
\usage{
f_stat_diff_of_means_medians(df, col_group, col_variable)
}
\arguments{
\item{df}{dataframe}

\item{col_group}{character vector denoting grouping variable}

\item{col_variable}{character vector denoting variable}
}
\value{
dataframe
}
\description{
used as a helper function for f_stat_anova
}
\examples{
set.seed(1)
df = tibble( fct = sample(LETTERS[1:5], 100, replace = T)
             , v1  = 1
             , v2  = rnorm(100, 4)
             , v3  = c( rep(3, 50), rep(8,50) )
)

col_group = 'fct'

f_stat_diff_of_means_medians(df, col_group, 'v1') \%>\%
  bind_rows( f_stat_diff_of_means_medians(df, col_group, 'v2') ) \%>\%
  bind_rows( f_stat_diff_of_means_medians(df, col_group, 'v3') )
}
