% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_group_ana_taglist}
\alias{f_stat_group_ana_taglist}
\title{analyse group difference of dataset}
\usage{
f_stat_group_ana_taglist(data_ls, col_group, tresh_p_val = 0.05,
  thresh_diff_perc = 3)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_group}{character vector denoting grouping columns}

\item{tresh_p_val}{p value threshold for plots, Default: 0.05}

\item{thresh_diff_perc}{minimum percent difference threshold for plots,
Default: 3}
}
\value{
taglist
}
\description{
returns a full analysis as a taglist inluding all features with
 p_values, medians, means, percentages and counts, as well as plots passing
 the treshold values
}
\examples{
\dontrun{
 data_ls = f_clean_data(mtcars)
 taglist = f_stat_group_ana_taglist(data_ls, 'cyl')
 f_plot_obj_2_html(taglist, type = "taglist", output_file = 'test_me', title = 'Plots')
 file.remove('test_me.html')
}
}
\seealso{
\code{\link[plotly]{ggplotly}}
 \code{\link[htmltools]{tagList}},\code{\link[htmltools]{h1}},\code{\link[htmltools]{h2}}
}
