% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_group_counts_percentages}
\alias{f_stat_group_counts_percentages}
\title{create a aggregated data frame with percentagers and counts for categorical variables}
\usage{
f_stat_group_counts_percentages(data_ls, col_group)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_group}{character vector denoting grouping columns}
}
\value{
dataframe
}
\examples{
f_clean_data( mtcars) \%>\%
 f_stat_group_counts_percentages('cyl')
}
