% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_max_diff_of_freq}
\alias{f_stat_max_diff_of_freq}
\title{calculate the maximal difference in frequencies between to categorical variables}
\usage{
f_stat_max_diff_of_freq(df, col_var1, col_var2)
}
\arguments{
\item{df}{dataframe containing both avariables}

\item{col_var1}{character vector denoting variable column 1}

\item{col_var2}{character vector denoting variable column 2}
}
\value{
dataframe
}
\description{
used as a helper function for f_stat_chi_square
}
\examples{
data_ls = f_clean_data(mtcars)
df_chi_squ = f_stat_max_diff_of_freq(data_ls$data, 'cyl', 'gear')
}
