% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_shapiro}
\alias{f_stat_shapiro}
\title{wrapper for shapiro.test()}
\usage{
f_stat_shapiro(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
shapiro.test object or list( statistic = NA, p.value = NA )
}
\description{
shapiro.test i slimited to <5000 sample size and raises an error
  if sd(x) == 0. Wrapper samples from input vector and returns a list object
  with NA parameters if sd(x) == 0.
}
\examples{
f_stat_shapiro( rnorm(1000, 10, 1) )
f_stat_shapiro( runif(1000, 1, 10) )
}
