% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_stars}
\alias{f_stat_stars}
\title{calculate significant level from p value}
\usage{
f_stat_stars(p_value)
}
\arguments{
\item{p_value}{numeric}
}
\value{
character vector
}
\description{
* P:0.05, ** P:0,005, *** P:0.001
}
\examples{
f_stat_stars(0.06)
f_stat_stars(0.05)
f_stat_stars(0.005)
f_stat_stars(0.001)
}
