% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot.R
\name{f_taglist_2_html}
\alias{f_taglist_2_html}
\title{generate a separate html file from a shiny taglist}
\usage{
f_taglist_2_html(taglist, output_file, title = "Plots")
}
\arguments{
\item{taglist}{htmltools::tagList}

\item{output_file}{file_name of the html file, without .html suffix}

\item{title}{character vector of html document title, Default: 'Plots'}
}
\description{
html objects generated by plotly or DT can be stored in
  htmltools::tagList. It can be printed in a html file to display the objects.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
