% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_train.R
\name{f_train_lasso_old}
\alias{f_train_lasso_old}
\title{wrapper for glmnet and HDtweedie}
\usage{
f_train_lasso_old(data, formula, grid = 10^seq(4, -4, length = 100),
  p = c(1, 1.25, 1.5, 1.75, 2), k = 5, family = "gaussian")
}
\arguments{
\item{data}{dataframe}

\item{formula}{formula}

\item{grid}{grid values for lambda, Default: 10^seq(4, -4, length = 100)}

\item{p}{p parameter for tweedie distributions, set p = NULL for not
performing lasso for tweedie distributions, Default: c(1, 1.25, 1.5, 1.75,
2)}

\item{k}{fold cross validation, set to 1 for testing against training data,
Default: 5}

\item{family}{family parameter for glmnet, can be a vector, Default:
'gaussian'}
}
\value{
list()
}
\description{
performs lasso for different distributions, returns a list of
  formulas that result in the lowest rtmse for at least one of the
  distributions. Graphical output allows side-by-side comparison of lasso
  behaviour for all distributions.
}
\details{
Columns containing NA will be removed, formula cannot be constructed
  with '.'
}
\examples{
data = MASS::quine
formula = Days~.

lasso = f_train_lasso(data, formula, p = NULL, k = 1
                     , grid = 10^seq(3,-3,length= 25) )
lasso = f_train_lasso(data, formula, p = 1.5, k = 2
                     , grid = 10^seq(3,-3,length= 25) )

lasso
}
\seealso{
\code{\link[HDtweedie]{HDtweedie}} \code{\link[glmnet]{glmnet}}
  \code{\link[pipelearner]{pipelearner}},\code{\link[pipelearner]{learn_models}},\code{\link[pipelearner]{learn_cvpairs}},\code{\link[pipelearner]{learn}}
}
