% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vignette.R
\name{f_vign_get_path}
\alias{f_vign_get_path}
\alias{f_vign_get_file_type}
\alias{f_vign_render}
\title{returns vignette path}
\usage{
f_vign_get_path()

f_vign_get_file_type(file_type = "html")

f_vign_render(overwrite = F)
}
\arguments{
\item{file_type}{character vector denoting file extensions}

\item{overwrite}{logical, Default: F}
}
\examples{
\dontrun{
 f_vign_get_path()
}
\dontrun{
 f_vign_get_file_type()
}
\dontrun{
 f_vign_render()
}
}
