% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vignette.R
\name{f_vignettes}
\alias{f_vignettes}
\title{open all vignettes}
\usage{
f_vignettes(vignettes = f_vign_get_file_type("html"), render_missing = T)
}
\arguments{
\item{vignettes}{vector with filenames of vignettes without file extensions, Default: f_vign_get_file_type("html")}

\item{render_missing}{logical, Default: T}
}
\description{
opens all oetteR vinettes in default browser, renders vignettes if they are not rendered yet.
}
\examples{
\dontrun{
 f_vignettes()
}
}
