% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats.R
\name{f_model_importance_plot}
\alias{f_model_importance_plot}
\title{plot model importance}
\usage{
f_model_importance_plot(importance, model_name, models.id = NULL,
  cv_pairs.id = NULL, train_p = NULL, variable_color_code = NULL, ...)
}
\arguments{
\item{importance}{dataframe importance created by f_model_importance()}

\item{model_name}{character vector (model column in pipelearner dataframe)
will be pasted for plot title}

\item{models.id}{character vector will be pasted for plot title}

\item{cv_pairs.id}{character vector will be pasted for plot title}

\item{train_p}{character vector will be pasted for plot title}

\item{variable_color_code}{dataframe created by f_plot_color_code_variables()}

\item{...}{additional character vectors to be pasted to plot title}
}
\value{
plotly graph
}
\description{
optimised for usage in pipelearner dataframe
}
\examples{

data_ls = f_clean_data(mtcars)
variable_color_code = f_plot_color_code_variables(data_ls)
m = twidlr::rpart(mtcars, disp~.)
imp = f_model_importance_rpart(m)
f_model_importance_plot(imp
                        , model_name = 'rpart'
                        , variable_color_code = variable_color_code
                        , 'example')
#pipelearner
pl = pipelearner::pipelearner(data_ls$data) \%>\%
  pipelearner::learn_models( twidlr::rpart, disp~. ) \%>\%
  pipelearner::learn_models( twidlr::randomForest, disp~. ) \%>\%
  pipelearner::learn_models( twidlr::svm, disp~. ) \%>\%
  pipelearner::learn() \%>\%
  mutate( imp = map2(fit, train, f_model_importance)
         , plot = pmap( list(imp, model, models.id, cv_pairs.id, train_p)
                       , f_model_importance_plot
                       , variable_color_code = variable_color_code
                       , 'pipelearner'
                       )
         )
htmltools::tagList(pl$plot)
}
