% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats_importance.R
\name{f_model_importance_plot_tableplot}
\alias{f_model_importance_plot_tableplot}
\title{tableplot of important variables}
\usage{
f_model_importance_plot_tableplot(data, ranked_variables, response_var,
  limit = 10, ...)
}
\arguments{
\item{data}{dataframe}

\item{ranked_variables}{datafram as returned by f_model_importance()}

\item{response_var}{character vector denoting response variable}

\item{limit}{integer limit the number of variables , Default: 10}

\item{...}{pass kwargs to tabplot::tableplot}
}
\value{
tabplot::tableplot object
}
\description{
takes the most important variables of a model and plots a tabplot::tableplot
}
\examples{

data = f_clean_data(mtcars) \%>\%
  .$data
m = rpart::rpart( disp~., data)
ranked_variables  = f_model_importance(m, data)
response_var = 'disp'

f_model_importance_plot_tableplot( data, ranked_variables, response_var, limit = 5 )

#pipe
form = as.formula('disp~cyl+mpg+hp')
pl = pipelearner::pipelearner(mtcars) \%>\%
 pipelearner::learn_models( twidlr::rpart, form ) \%>\%
 pipelearner::learn_models( twidlr::randomForest, form ) \%>\%
 pipelearner::learn_models( twidlr::svm, form ) \%>\%
 pipelearner::learn() \%>\%
 mutate( imp = map2(fit, train, f_model_importance)
         , tabplot = pmap( list( data = train
                                 , ranked_variables = imp
                                 , response_var = target
                                 , title = model
         )
         , f_model_importance_plot_tableplot
         , limit = 5
         )
  )
}
\seealso{
\code{\link[tabplot]{tableplot}}
}
