% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_model_feats_importance.R
\name{f_model_importance_svm}
\alias{f_model_importance_svm}
\title{extract variable importance for svm}
\usage{
f_model_importance_svm(m, data)
}
\arguments{
\item{m}{model of class svm}

\item{data}{original training dataframe}
}
\value{
dataframe
}
\details{
uses 1D-SA 1 dimensional sensitivity analysis using
  rminer::Importance()
}
\examples{

#regression
m = twidlr::svm(mtcars, disp~.)
f_model_importance_svm(m, mtcars)

#classification
data = mtcars
data$cyl = factor(data$cyl, ordered = T)
m = twidlr::svm(data, cyl~.)
f_model_importance_svm(m, data)


}
