% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_pca.R
\name{f_pca_get_plot_variance_explained}
\alias{f_pca_get_plot_variance_explained}
\title{FUNCTION_TITLE}
\usage{
f_pca_get_plot_variance_explained(pca_ls)
}
\arguments{
\item{pca_ls}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
