% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_pca.R
\name{f_pca_plot_components}
\alias{f_pca_plot_components}
\title{plot principle components as a dot plot}
\usage{
f_pca_plot_components(pca_ls, x_axis = "PC1", y_axis = "PC2",
  group = NULL)
}
\arguments{
\item{pca_ls}{list created by f_pca()}

\item{x_axis}{character vector, Default: 'PC1'}

\item{y_axis}{character vector, Default: 'PC2'}

\item{group}{character vector denoting the grouping variable, determines dot
colour, Default: NULL}
}
\value{
htmltools taglist containing a plotly graph and tow DT datatables
  will only show if printed in a rmarkdown document
}
\examples{
\dontrun{
tagls = f_clean_data(mtcars) \%>\%
  f_boxcox() \%>\%
  f_pca() \%>\%
  f_pca_plot_components(group = 'cyl')
}
}
