% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_plot.R
\name{f_plot_hist}
\alias{f_plot_hist}
\title{Plot Histograms}
\usage{
f_plot_hist(variable, data_ls, group = "None", graph_type = c("violin",
  "bar", "line"), y_axis = c("count", "density"), auto_range = T,
  n_breaks = 30, x_min = 0, x_max = 100)
}
\arguments{
\item{variable}{PARAM_DESCRIPTION}

\item{data_ls}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{graph_type}{PARAM_DESCRIPTION, Default: c("violin", "bar", "line")}

\item{y_axis}{PARAM_DESCRIPTION, Default: c("count", "density")}

\item{auto_range}{PARAM_DESCRIPTION, Default: T}

\item{n_breaks}{PARAM_DESCRIPTION, Default: 30}

\item{x_min}{PARAM_DESCRIPTION, Default: 0}

\item{x_max}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function plots smart histograms for variables in a data_ls list
 generated by f_clean_data(). It supports three types of histograms: Bar
 histograms, density histograms and violin plots. We can further specify a
 categorical variable to group on. The function defaults to a sensible
 standard output if key word arguments are not applicable for variable type.
 Thus we can easily pipe through long lists of variables and thus generate
 histograms for all variables in the input (see examples).
}
\details{
DETAILS
}
\examples{
\dontrun{#'
#plot single variable
data_ls = f_clean_data(mtcars)
f_plot_hist('disp', data_ls)
f_plot_hist('disp', data_ls, graph_type = 'violin', group = 'cyl')

#plot all variables

vars = c(data_ls$categorical, data_ls$numericals) \%>\%
 walk( f_plot_hist, data_ls, group = 'cyl')
}
}
\seealso{
\code{\link[stringr]{str_extract_all}},\code{\link[stringr]{str_c}}
}
