% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-f_plot.R
\name{f_plot_time}
\alias{f_plot_time}
\title{FUNCTION_TITLE}
\usage{
f_plot_time(variable, time_variable, data_ls, time_variable_as_factor = F,
  group = NULL, normalize = F, time_unit = "day")
}
\arguments{
\item{variable}{PARAM_DESCRIPTION}

\item{time_variable}{PARAM_DESCRIPTION}

\item{data_ls}{PARAM_DESCRIPTION}

\item{time_variable_as_factor}{PARAM_DESCRIPTION, Default: F}

\item{group}{PARAM_DESCRIPTION, Default: NULL}

\item{normalize}{PARAM_DESCRIPTION, Default: F}

\item{time_unit}{PARAM_DESCRIPTION, Default: 'day'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
data       = nycflights13::flights
data$is_ua = ifelse( data$carrier == 'UA', 'UA', 'other')
data$date  = data$year * 10000 + data$month * 100 + data$day
data$date  = lubridate::as_date( data$date )
data_ls    = f_clean_data( data, replace_neg_values_with_zero = F)
f_plot_time( 'arr_delay', 'month', data_ls, group = 'is_ua', time_unit = 'month')

}
}
