% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_pl_regression_summarize}
\alias{f_predict_pl_regression_summarize}
\title{summarize prediction by f_predict_pl_regression()}
\usage{
f_predict_pl_regression_summarize(pl)
}
\arguments{
\item{pl}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link{f_predict_pl_regression}}
}
