% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_plot_model_performance_regression}
\alias{f_predict_plot_model_performance_regression}
\title{plot model performance}
\usage{
f_predict_plot_model_performance_regression(data)
}
\arguments{
\item{data}{dataframe with the columns title, bins, resid_abs, resid_squ, ape}
}
\value{
taglist \describe{
  \item{[1]}{Headline summary Plots }
  \item{[2]}{Residuals Pointplot}
  \item{[3]}{Residuals Boxplot}
  \item{[4]}{APE Pointplot}
  \item{[5]}{APE Boxplot}
  \item{[6]}{MAPE, MSE, MAE, Binning }
  \item{[7]}{Headline Performance Measures Summary}
  \item{[8]}{Summary MAPE, MSE, MAE with SE }
  \item{[9]}{Summary MAPE, MSE, MAE with CI95 }
  \item{[10]}{Headline Summary Tables}
  \item{[11]}{Summary Table}
  \item{[12]}{Table Binning}
}
}
\description{
add predictions to modelling dataframe and unnest, create a title column and a bins column
}
\examples{
\dontrun{
form = as.formula( 'displacement~cylinders+mpg')

ISLR::Auto \%>\%
 pipelearner::pipelearner() \%>\%
 pipelearner::learn_models( twidlr::rpart, form ) \%>\%
 pipelearner::learn_models( twidlr::randomForest, form ) \%>\%
 pipelearner::learn_models( twidlr::svm, form ) \%>\%
 pipelearner::learn() \%>\%
 f_predict_pl_regression( 'name' ) \%>\%
 unnest(preds) \%>\%
 mutate( bins = cut(target1, breaks = 3 , dig.lab = 4)
         , title = paste(models.id, cv_pairs.id, train_p, target, model) ) \%>\%
 f_predict_plot_model_performance_regression() \%>\%
 f_plot_obj_2_html(type = 'taglist', 'test_me', title = 'Model Performance')

file.remove('test_me.html')
}
}
\seealso{
\code{\link[htmltools]{tagList}}
 \code{\link[plotly]{ggplotly}}
 \code{\link[DT]{datatable}}
 \code{\link{f_predict_pl_regression} }
}
