% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_predict.R
\name{f_predict_regression_add_predictions}
\alias{f_predict_regression_add_predictions}
\title{adds predictions, residuals, abolute residuals, squared residuals and
  absolute percent error to a dataframe.}
\usage{
f_predict_regression_add_predictions(data, m, col_target, cols_id = NULL,
  formula = NULL, ...)
}
\arguments{
\item{m}{regression model}

\item{col_target}{character vector naming target/response variable}

\item{cols_id}{character vector naming id columns, if specified non_id
columns will be dropped from dataframe, in order to be more memory
efficient.}

\item{df}{dataframe containing data to be used as the basis for prediction.
Can also be a modelR resample object}
}
\value{
dataframe
}
\description{
absolute percent error = (abs(resid/pred)*100 )
}
\details{
works with HDtweedie, randomForest, rpart, e1071::svm
}
\examples{
df = mtcars \%>\%
mutate(names = row.names(.))
m = rpart::rpart(disp~., df)
pred = f_predict_regression_add_predictions(df, m, 'disp', 'names')
pred
}
\seealso{
\code{\link[modelr]{add_predictions}},\code{\link[modelr]{add_residuals}}
}
