% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sim.R
\name{f_sim_profit}
\alias{f_sim_profit}
\title{simulate profit}
\usage{
f_sim_profit(output_file = "profit_simulation", path = ".",
  params = "ask")
}
\arguments{
\item{output_file}{PARAM_DESCRIPTION, Default: 'profit_simulation'}

\item{path}{PARAM_DESCRIPTION, Default: '.'}

\item{params}{list params passed to rmarkdown::render(), see example, Default: 'ask'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
using the following parameters:
 \itemize{
 \item retention rate (retention_rate)
 \item retention common for business (retention_rate_common)
 \item new customers acquired per year (nca_per_year)
 \item expected increase in customers acquired  (expected_increase_nca)
 \item present number of customers (n_customers)
 \item fixed cost (fix_cost)
 \item profit per customer per year (profit_cm1_per_customer)
 }
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 f_sim_profit( path = tempdir() )
 f_sim_profit( prefix, params = list( retention_rate = 0.82
                                     , retention_rate_common = 0.88
                                     , nca_per_year = 4000
                                     , expected_increase_nca = 2
                                     , n_customers = 150000
                                     , fix_cost = 20000000
                                     , profit_cm1_per_customer = 200 )
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
