% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_stat.R
\name{f_stat_group_ana}
\alias{f_stat_group_ana}
\title{analyse group difference of dataset}
\usage{
f_stat_group_ana(data_ls, col_group, thresh_p_val = 0.05,
  thresh_diff_perc = 3, output_file = "group_ana", static_plots = T,
  alluvial = T, alluvial_thresh_p_val = 0.05,
  alluvial_thres_diff_perc = 7.5, tabplot = T, return_taglist = F)
}
\arguments{
\item{data_ls}{data_ls object generated by f_clean_data(), or a named list
list( data = <dataframe>, numericals = < vector with column names of
numerical columns>)}

\item{col_group}{character vector denoting grouping columns}

\item{thresh_p_val}{p value threshold for plots, Default: 0.05}

\item{thresh_diff_perc}{minimum percent difference threshold for plots,
Default: 3}

\item{output_file}{character vector containing output file name}
}
\value{
file path to html file
}
\description{
creates a html document with a group analysis
}
\examples{
\dontrun{
 data_ls = f_clean_data(mtcars)
 f_stat_group_ana(data_ls, 'cyl', output_file = 'test_me')
 file.remove('test_me.html')
 file.remove('test_me_stat_plots.html')
 file.remove('test_me_alluvial.html')
 file.remove('test_me_tabplots.html')
}
}
\seealso{
\code{\link[plotly]{ggplotly}}
 \code{\link[htmltools]{tagList}},\code{\link[htmltools]{h1}},\code{\link[htmltools]{h2}}
}
